Add class to count active slides
This adds an additional class to each active slide in the form of `tns-slide-active-[nth]` where `[nth]` is the counted up from the first active slide.

E.g., if 3 slides are active, they will be assigned the following classes respectively:
```
.tns-slide-active-0
.tns-slide-active-1
.tns-slide-active-2
```

This allows them to be styled. Differently. For instance, the first and last slides can be blurred as requested in https://github.com/ganlanyuan/tiny-slider/issues/586.